Convert picker to reusable component
